+++ src/IdBrokerClient.php
@@ -23,7 +23,7 @@ class IdBrokerClient extends BaseClient
      *
      * @var IPBlock[]
      */
-    private $trustedIpRanges = [];
+    private $trustedIpRanges = [ ];
 
     private $assertValidBrokerIp = true;
 
@@ -36,11 +36,13 @@ class IdBrokerClient extends BaseClient
      *     Example: 'https://api.example.com/'.
      * @param string $accessToken - Your authorization access (bearer) token.
      * @param array $config - Any other configuration settings.
+     * @throws \InvalidArgumentException
+     * @throws \Exception
      */
     public function __construct(
         string $baseUri,
         string $accessToken,
-        array $config = []
+        array $config = [ ]
     ) {
         if (empty($baseUri)) {
             throw new \InvalidArgumentException(
@@ -75,20 +77,22 @@ class IdBrokerClient extends BaseClient
         ], $config));
     }
 
-    /*
+    /**
      * Validates the config values for ASSERT_VALID_BROKER_IP_CONFIG and
-     *   ASSERT_VALID_BROKER_IP_CONFIG
+     *   TRUSTED_IPS_CONFIG.
      * Uses them to set $this->assertValidBrokerIp and $this->trustedIpRanges
      *
      * @param array the config values for the client
      *
      * @return null
      * @throws \InvalidArgumentException
+     * @throws \Exception if assertValidBrokerIp is true and idBrokerUri is invalid, unresolvable, or untrusted
      */
-    private function initializeConfig($config) {
+    private function initializeConfig($config)
+    {
 
-        if ( isset($config[self::ASSERT_VALID_BROKER_IP_CONFIG])) {
-            $this->assertValidBrokerIp = $config[self::ASSERT_VALID_BROKER_IP_CONFIG];
+        if (isset($config[ self::ASSERT_VALID_BROKER_IP_CONFIG ])) {
+            $this->assertValidBrokerIp = $config[ self::ASSERT_VALID_BROKER_IP_CONFIG ];
         }
 
         // If we don't need to validate the Broker Ip, we're done here
@@ -100,7 +104,7 @@ class IdBrokerClient extends BaseClient
          *  If we should validate the Broker IP but there aren't
          *  any trusted IPs, throw an exception
          */
-        if (empty($config[self::TRUSTED_IPS_CONFIG])) {
+        if (empty($config[ self::TRUSTED_IPS_CONFIG ])) {
             throw new \InvalidArgumentException(
                 'The config entry for ' . self::TRUSTED_IPS_CONFIG .
                 ' must be set (as an array) when ' .
@@ -114,7 +118,7 @@ class IdBrokerClient extends BaseClient
          * At this point, we need to validate the Broker Ip and we know
          * that the TRUSTED_IPS_CONFIG is not empty
          */
-        $newTrustedIpRanges = $config[self::TRUSTED_IPS_CONFIG];
+        $newTrustedIpRanges = $config[ self::TRUSTED_IPS_CONFIG ];
         if ( ! is_array($newTrustedIpRanges)) {
             throw new \InvalidArgumentException(
                 'The config entry for ' . self::TRUSTED_IPS_CONFIG .
@@ -125,7 +129,7 @@ class IdBrokerClient extends BaseClient
 
         foreach ($newTrustedIpRanges as $nextIpRange) {
             $ipBlock = IPBlock::create($nextIpRange);
-            $this->trustedIpRanges[] = $ipBlock;
+            $this->trustedIpRanges[ ] = $ipBlock;
         }
 
         $this->assertTrustedBrokerIp();
@@ -139,7 +143,7 @@ class IdBrokerClient extends BaseClient
      * @param string $username The username.
      * @param string $password The password (in plaintext).
      * @return array|null An array of user information (if valid), or null.
-     * @throws Exception
+     * @throws ServiceException
      */
     public function authenticate(string $username, string $password)
     {
@@ -147,7 +151,7 @@ class IdBrokerClient extends BaseClient
             'username' => $username,
             'password' => $password,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
@@ -155,34 +159,57 @@ class IdBrokerClient extends BaseClient
             return null;
         }
         
-        $this->reportUnexpectedResponse($result, 1490802360, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490802360);
     }
-    
+
+    /**
+     * Attempt to authenticate using a new user invite code
+     *
+     * @param string $invite The new user invite code.
+     * @return array|null An array of user information (if valid), or null.
+     * @throws ServiceException
+     */
+    public function authenticateNewUser(string $invite)
+    {
+        $result = $this->authenticateNewUserInternal([
+            'invite' => $invite,
+        ]);
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        } elseif ($statusCode === 400) {
+            return null;
+        }
+
+        $this->reportUnexpectedResponse($result, 1544549972);
+    }
+
     /**
      * Create a user with the given information.
      *
      * @param array $config An array key/value pairs of attributes for the new
      *     user.
      * @return array An array of information about the new user.
-     * @throws Exception
+     * @throws ServiceException
      */
-    public function createUser(array $config = [])
+    public function createUser(array $config = [ ])
     {
         $result = $this->createUserInternal($config);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
         
-        $this->reportUnexpectedResponse($result, 1490802526, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490802526);
     }
     
     /**
      * Deactivate a user.
      *
      * @param string $employeeId The Employee ID of the user to deactivate.
-     * @throws Exception
+     * @throws ServiceException
      */
     public function deactivateUser(string $employeeId)
     {
@@ -190,10 +217,10 @@ class IdBrokerClient extends BaseClient
             'employee_id' => $employeeId,
             'active' => 'no',
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode !== 200) {
-            $this->reportUnexpectedResponse($result, 1490808523, $statusCode);
+            $this->reportUnexpectedResponse($result, 1490808523);
         }
     }
     
@@ -205,7 +232,7 @@ class IdBrokerClient extends BaseClient
      */
     protected function getResultAsArrayWithoutStatusCode($result)
     {
-        unset($result['statusCode']);
+        unset($result[ 'statusCode' ]);
         return $result->toArray();
     }
 
@@ -213,18 +240,18 @@ class IdBrokerClient extends BaseClient
      * Ping the /site/status url
      *
      * @return string "OK".
-     * @throws Exception
+     * @throws ServiceException
      */
     public function getSiteStatus()
     {
         $result = $this->getSiteStatusInternal();
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if (($statusCode >= 200) && ($statusCode < 300)) {
             return 'OK';
         }
 
-        $this->reportUnexpectedResponse($result, 1490806100, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490806100);
     }
     
     /**
@@ -233,14 +260,14 @@ class IdBrokerClient extends BaseClient
      * @param string $employeeId The Employee ID of the desired user.
      * @return array|null An array of information about the specified user, or
      *     null if no such user was found.
-     * @throws Exception
+     * @throws ServiceException
      */
     public function getUser(string $employeeId)
     {
         $result = $this->getUserInternal([
             'employee_id' => $employeeId,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
@@ -248,7 +275,7 @@ class IdBrokerClient extends BaseClient
             return null;
         }
         
-        $this->reportUnexpectedResponse($result, 1490808555, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490808555);
     }
     
     /**
@@ -259,22 +286,23 @@ class IdBrokerClient extends BaseClient
      * @param  array|null $search (Optional:) An array of fields to search on,
      *     example ['username' => 'billy']
      * @return array An array with a sub-array about each user.
+     * @throws ServiceException
      */
-    public function listUsers($fields = null, $search = [])
+    public function listUsers($fields = null, $search = [ ])
     {
-        $config = [];
+        $config = [ ];
         if ($fields !== null) {
-            $config['fields'] = join(',', $fields);
+            $config[ 'fields' ] = join(',', $fields);
         }
         $config = array_merge($config, $search);
         $result = $this->listUsersInternal($config);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
         
-        $this->reportUnexpectedResponse($result, 1490808715, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490808715);
     }
 
     /**
@@ -283,7 +311,7 @@ class IdBrokerClient extends BaseClient
      * @param string $type
      * @param string $label
      * @return array|null
-     * @throws Exception
+     * @throws ServiceException
      */
     public function mfaCreate($employee_id, $type, $label = null)
     {
@@ -292,20 +320,21 @@ class IdBrokerClient extends BaseClient
             'type' => $type,
             'label' => $label,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
 
-        $this->reportUnexpectedResponse($result, 1506710701, $statusCode);
+        $this->reportUnexpectedResponse($result, 1506710701);
     }
 
     /**
      * Delete an MFA configuration
      * @param int $id
+     * @param string $employeeId
      * @return null
-     * @throws Exception
+     * @throws ServiceException
      */
     public function mfaDelete($id, $employeeId)
     {
@@ -313,32 +342,57 @@ class IdBrokerClient extends BaseClient
             'id' => $id,
             'employee_id' => $employeeId,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if ($statusCode === 204) {
             return null;
         }
 
-        $this->reportUnexpectedResponse($result, 1506710702, $statusCode);
+        $this->reportUnexpectedResponse($result, 1506710702);
     }
 
     /**
      * Get a list of MFA configurations for given user
      * @param string $employee_id
      * @return array
+     * @throws ServiceException
      */
     public function mfaList($employee_id)
     {
         $result = $this->mfaListInternal([
             'employee_id' => $employee_id,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
 
-        $this->reportUnexpectedResponse($result, 1506710703, $statusCode);
+        $this->reportUnexpectedResponse($result, 1506710703);
+    }
+
+    /**
+     * Update an MFA configuration
+     * @param int $id
+     * @param string $employeeId
+     * @param string $label
+     * @return array
+     * @throws ServiceException
+     */
+    public function mfaUpdate($id, $employeeId, $label)
+    {
+        $result = $this->mfaUpdateInternal([
+            'id' => $id,
+            'employee_id' => $employeeId,
+            'label' => $label,
+        ]);
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        }
+
+        $this->reportUnexpectedResponse($result, 1543879805);
     }
 
     /**
@@ -348,6 +402,7 @@ class IdBrokerClient extends BaseClient
      * @param string $value The MFA value being verified.
      * @return bool
      * @throws MfaRateLimitException
+     * @throws ServiceException
      */
     public function mfaVerify($id, $employeeId, $value)
     {
@@ -356,7 +411,7 @@ class IdBrokerClient extends BaseClient
             'employee_id' => $employeeId,
             'value' => $value,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if ($statusCode === 204) {
             return true;
@@ -366,7 +421,127 @@ class IdBrokerClient extends BaseClient
             throw new MfaRateLimitException('Too many recent failures for this MFA');
         }
 
-        $this->reportUnexpectedResponse($result, 1506710704, $statusCode);
+        $this->reportUnexpectedResponse($result, 1506710704);
+    }
+
+    /**
+     * Create a new recovery method
+     * @param string $employee_id
+     * @param string $value
+     * @param string $created If specified, indicates the record is to be created pre-verified.
+     * @return String[]
+     * @throws ServiceException
+     */
+    public function createMethod($employee_id, $value, $created = '')
+    {
+        $params = compact('employee_id', 'value');
+        if (! empty($created)) {
+            $params['created'] = $created;
+        }
+
+        $result = $this->createMethodInternal($params);
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006274);
+    }
+
+    /**
+     * Delete a recovery method
+     * @param int $uid
+     * @param int $employee_id
+     * @return null
+     * @throws ServiceException
+     */
+    public function deleteMethod($uid, $employee_id)
+    {
+        $result = $this->deleteMethodInternal(compact('uid', 'employee_id'));
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 204 || $statusCode === 200) {
+            return null;
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006315);
+    }
+
+    /**
+     * View a single recovery method
+     * @param int $uid
+     * @param int $employee_id
+     * @return String[]
+     * @throws ServiceException
+     */
+    public function getMethod($uid, $employee_id)
+    {
+        $result = $this->getMethodInternal(compact('uid', 'employee_id'));
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006615);
+    }
+
+    /**
+     * Get a list of recovery methods for given user
+     * @param String $employee_id
+     * @return String[]
+     * @throws ServiceException
+     */
+    public function listMethod($employee_id)
+    {
+        $result = $this->listMethodInternal(compact('employee_id'));
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006346);
+    }
+
+    /**
+     * Verify a recovery method
+     * @param string $uid The Method UID.
+     * @param string $employee_id The Employee ID of the user with that Method.
+     * @param string code The recovery method verification code
+     * @return String[]
+     * @throws ServiceException
+     */
+    public function verifyMethod($uid, $employee_id, $code)
+    {
+        $result = $this->verifyMethodInternal(compact('uid', 'employee_id', 'code'));
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 200) {
+            return $this->getResultAsArrayWithoutStatusCode($result);
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006448);
+    }
+
+    /**
+     * Resend a recovery method verification message
+     * @param string $uid The Method UID.
+     * @param string $employee_id The Employee ID of the user with that Method.
+     * @return bool
+     * @throws ServiceException
+     */
+    public function resendMethod($uid, $employee_id)
+    {
+        $result = $this->resendMethodInternal(compact('uid', 'employee_id'));
+        $statusCode = (int)$result[ 'statusCode' ];
+
+        if ($statusCode === 204 || $statusCode === 200) {
+            return true;
+        }
+
+        $this->reportUnexpectedResponse($result, 1541006732);
     }
 
     /**
@@ -377,7 +552,7 @@ class IdBrokerClient extends BaseClient
      * @param string $password The desired (new) password, in plaintext.
      *
      * @return array An array of password metadata
-     * @throws Exception
+     * @throws ServiceException
      */
     public function setPassword(string $employeeId, string $password)
     {
@@ -385,16 +560,21 @@ class IdBrokerClient extends BaseClient
             'employee_id' => $employeeId,
             'password' => $password,
         ]);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
 
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
 
-        $this->reportUnexpectedResponse($result, 1490808839, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490808839);
     }
-    
-    protected function reportUnexpectedResponse($response, $uniqueErrorCode, $httpStatusCode = null)
+
+    /**
+     * @param \GuzzleHttp\Command\Result $response
+     * @param int $uniqueErrorCode
+     * @throws ServiceException
+     */
+    protected function reportUnexpectedResponse($response, $uniqueErrorCode)
     {
         throw new ServiceException(
             sprintf(
@@ -402,43 +582,44 @@ class IdBrokerClient extends BaseClient
                 var_export($response, true)
             ),
             $uniqueErrorCode,
-            $httpStatusCode
+            (int)$response[ 'statusCode' ]
         );
     }
-    
+
     /**
      * Update the specified user with the given information.
      *
      * @param array $config An array key/value pairs of attributes for the user.
      *     Must include at least an 'employee_id' entry.
      * @return array An array of information about the updated user.
-     * @throws Exception
+     * @throws ServiceException
      */
-    public function updateUser(array $config = [])
+    public function updateUser(array $config = [ ])
     {
         $result = $this->updateUserInternal($config);
-        $statusCode = (int)$result['statusCode'];
+        $statusCode = (int)$result[ 'statusCode' ];
         
         if ($statusCode === 200) {
             return $this->getResultAsArrayWithoutStatusCode($result);
         }
         
-        $this->reportUnexpectedResponse($result, 1490808841, $statusCode);
+        $this->reportUnexpectedResponse($result, 1490808841);
     }
 
     /**
      * Determine whether any of the Id-broker's IPs are not in the
      * trusted ranges
      *
-     * @throws Exception
+     * @throws Exception if idBrokerUri is invalid, unresolvable, or untrusted
      */
-    private function assertTrustedBrokerIp() {
+    private function assertTrustedBrokerIp()
+    {
         $baseHost = parse_url($this->idBrokerUri, PHP_URL_HOST);
         $idBrokerIp = gethostbyname(
             $baseHost
         );
 
-        if ( ! $this->isTrustedIpAddress($idBrokerIp)) {
+        if (! $this->isTrustedIpAddress($idBrokerIp)) {
             throw new Exception(
                 'The Id Broker has an IP that is not trusted ... ' . $idBrokerIp,
                 1494531300